+++ src/main/java/com/enrique/reservatusalaback/repository/CustomerRepository.java
@@ -3,5 +3,6 @@ package com.enrique.reservatusalaback.repository;
 import com.enrique.reservatusalaback.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+
 }
