+++ src/main/java/com/enrique/reservatusalaback/repository/RoomRepository.java
@@ -3,6 +3,6 @@ package com.enrique.reservatusalaback.repository;
 import com.enrique.reservatusalaback.model.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
 
 }
